requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
flake8==4.0.1
selenium==3.141.0
requests==2.28.1
webdriver-manager==3.8.5
fake-useragent==1.1.1
